Change default for proto_library.cc_lib
The default for cc_lib should depend on whether or not you want to force the lite runtime. This changes it to a target that takes the feature configuration into account.

RELNOTES=Change the default for cc_lib based to take into account if the lite proto runtime is being forced.
PiperOrigin-RevId: 211636197