Rewrite the BuildEventServiceTransport
The code of the BuildEventServiceTransport has become sufficiently
complex, difficult to reason about and hard to test that we found
it necessary to rewrite it before launching this in production.

The new implementation uses an eventloop style concurrency model
for build event uploads. Additionally, the BuildEventServiceClient
implementations where updated to only report errors via the Future
returned by BESC.openStream(...). The BuildEventServiceProtoUtil
was immutable and is no longer responsible for assigning sequence
numbers.

RELNOTES: None
PiperOrigin-RevId: 211784727