Lazy output stream for test.xml, does not produce a file if no test xml data was produced
This fixes #5289.

The general problem is, if an exception is thrown during creation of the `TestSuiteModel`, the test runner will exit, closing off the OutputStream it had created for writing the test.xml. Unfortunately, if nothing was actually written to the stream, closing it will create a 0-byte test.xml on disk regardless.

This PR wraps the OutputStream in a lazy producer, only creating the actual stream if something was written to it. This guarantees test.xml will contain actual content.

Closes #6000.

PiperOrigin-RevId: 211799139