Automated rollback of commit 395fbbdb7f17b97ed140fc0e4a1e061c7bf73843.
*** Reason for rollback ***

Breaking blaze guitar tests

*** Original change description ***

Allow shadowing of builtins in bzl files

This was previously allowed only with --incompatible_static_name_resolution (although it was a backward-compatible change). Removing the code path simplifies the code and makes easier to implement the static name resolution proposal.

The downside is that it will temporarily allow this corner-case:

```
x = len(...) # reference to the builtin
len = ...    # shadow the builtin with a new global
```

RELNOTES: None.
PiperOrigin-RevId: 212017164