Allow passing list into user_compile_flags and user_link_flags
Depset has advantage that it doesn't grow quadratically in memory, but a
disadvantage, that it removes duplicates. Which is especially bad when
user_compile_flags contains [ "-isystem", "a", "-isystem", "b" ] as it will
result in [ "-isystem", "a", "b" ] passed to the compiler, and the compiler
won't be happy.

RELNOTES: None.
PiperOrigin-RevId: 212223261