C++: Stops calling link() from CcImport
There is no reason to call link() in CcImport.java, we can build a
CcLinkingInfo provider directly. The logic is largely borrowed from the Skylark
implementation.

The Skylark implementation will be the only implementation in the future, but
for now in order to make other refactorings in the C++ rules easier, we remove
the call to link() here.

RELNOTES:none
PiperOrigin-RevId: 212792371