Don't attempt to evaluate the top-level transition if the target's in error
A transition may want to throw an unchecked exception to fail-fast if it's used in a RuleClass that doesn't set the right preconditions. For example, the transition might allow only certain attribute values, and want to alert the RuleClass author asap if they forgot to set Attribute.Builder#allowedValues.

If the RuleClass's preconditions are working correctly, then the unchecked exception will be unreachable during normal builds, because we would never get as far as evaluating the transition if the target is in error. This CL makes sure the same is true of our BuildViewTestCase tests.

This is needed in order to easily assert that a target fails validation, without the test evaluation being interrupted by an unnecessary exception.

RELNOTES: None
PiperOrigin-RevId: 212808128