C++: Breaks up C++ linking API.
The API as it is right now is not flexible. It is written to match cc_library's
use case without taking into account many of the other usages. This forces us
to add a lot of booleans for switching logic. The bits that are broken apart
and used by more than one caller can still be shared but not as part of the
link() API method.

This will make it easier to add support for the linking necessary for
implementing cc_binary and transitive linking in general.

RELNOTES:none
PiperOrigin-RevId: 212952298