C++: Another refactoring for CcLinkingHelper.
The CcNativeLibraryProvider is specific to cc_library and C++ proto rules. This shouldn't be in the API and is pulled out to the respective callsites.

In addition, linking libraries when there are no object files is only ever done from cc_library. This was another boolean in CcLinkingHelper that is removed.

RELNOTES:none
PiperOrigin-RevId: 212961089