Replace FileValue with ArtifactFileMetadata for tracking metadata during execution. FileValue unnecessarily keeps a reference to a Path (via RootedPath) which is not needed for execution.
Did a drive-by functional change of getting rid of special-casing for empty file in ActionMetadataHandler: special handling of empty files has been gone for years. Also reworked comments to reflect the current state of the world (namely, additionalOutputData is often the primary source of metadata).

Only potential semantic change I can see is that we now get the last modified time of an artifact in FilesystemValueChecker by doing a getLastModifiedTime on the artifact's path directly, versus the real path of the FileValue. That is safe, though, because the real path of the FileValue came directly from our just reading the artifact this time.

PiperOrigin-RevId: 213005841