external repositories: support scaling timeouts
Authors of repository rules can provide a timeout for commands executed
as part of this rule. However, there is no value that universally fits:
the time needed for a `git clone` depends on the size of the (shallow)
clone requested as well as the network characteristics to the repository.

Therefore, add an option to scale all timeouts in repository rules by
a given factor. In this way, an external repository can be made working
on a given (slow or badly connected) machine without having to change
the underlying source code.

Fixes #6110.

Change-Id: I856a46d6ab2950f52337240e4e2a43a98ece009e
PiperOrigin-RevId: 213769242