Remove the unnecessary unbounded thread pool
Even if we wanted to do the stats in a thread pool or async, we'd have
to do it differently. But it's probably not worth it - the stat results
should already be in the kernel cache (since we just stat the files
earlier - all action outputs are stat at the end of action execution).

PiperOrigin-RevId: 214400653