Include a stack trace showing what requested failed invalidations
Logs indicate that, sometimes, an invalidation is happening after an
evaluation that failed with an OOM. This invalidation fails because
the OOM breaks graph invariants. Annoyingly, the invariant breakage gets
reported instead of the OOM.

Wrapping an invalidation failure adds a stack trace which will show
exactly what's requesting the invalidation, which will suggest why the
evaluation's OOM isn't propagated.

RELNOTES: None.
PiperOrigin-RevId: 214463849