Make ExecutionInfoModifier have a useful equals by making it AutoValue.
All objects which are to be part of the build options must parse equal on repeated
invocations, or the analysis cache will be discarded unnecessarily.

It's also now possible to create an empty ExecutionInfoModifier. This is used
instead of null as the default value. This allows for more easily turning
--modify_execution_info off, even when it's in the RC file. (It's already
possible to do this by creating a silly regex like "(?!.*)=-nonsense-value",
so this isn't a new capability, it's just a more discoverable one.)

Also cleans up the parsing code to use Guava string-handling methods and
fixes some uses of plain non-linked HashMap surrounding the use of this.
Most users of this method were already using LinkedHashMaps, so this makes
callers more consistent.

RELNOTES: None.
PiperOrigin-RevId: 214622440