Remove unnecessary AttributeContainer bits
setAttributeLocation(Attribute, Location) is only used in tests and doesn't add
much on top of the public setAttributeLocation(int, Location) API. Delete it
and inline the attribute index getting logic in the tests.

Took the liberty of refactoring the tests a bit following me breaking them and
trying to understand why.

PiperOrigin-RevId: 214659408