C++: Make cc_binary's linking go through CcLinkingHelper.
CcLinkingHelper has new functionality to support transitive linking. This is
not exposed to Starlark yet.

GitHub issue #4570

This was rolled back because in the refactoring I started passing headers to the linking action that were not passed before. These headers have mistakes and cannot be pre-processed, so when they were passed as inputs to the linking action, the compilation action was executed whereas before it wasn't.

Fixed and added test to make sure the behavior is the same as before.

RELNOTES:none
PiperOrigin-RevId: 214785687