Automated rollback of commit c004ff117ad72fd387102600e4be47fddf7bc2ae.
*** Reason for rollback ***

Second attempt; rollout with a flag this time.

The original change should have been a no-op, but twerth@ swears that
his binary search narrowed down to it, so we're being super extra
careful this time, having been unable to find the root cause to the
issue that caused the rollback.

*** Original change description ***

Remove PerActionFileCache

Instead, make ActionMetadataHandler implement the MetadataProvider interface.

This fixes an issue where an action that runs two spawns where one depends on
an output of the other was unable to get the metadata for the intermediate
output.

We don't currently have actions that do this, but we will have in a future
change (which will also implicitly act as a regression test).

PiperOrigin-RevId: 214919699