Mark nodes to be force-rebuilt during node restarting
Non-deterministic nodes may evaluate to different values after being
restarted. By ensuring that the nodes are "force-rebuilt" after a
restart, such a situation won't result in a crash.

Because restarted nodes are no longer dirtied with isChanged=true, the
assertion logic that ensures that nodes are not dirtied twice with the
same value for isChanged can go back into the #markDirty method. In
other words, this CL includes a logical rollback of "Permit marking
dirty/changed a node more than once".

This also fixes a race condition bug where a dep which is restarted by
>1 parent could cause a crash if the dep was inflight when the later
restart attempted to dirty it.

RELNOTES: None.
PiperOrigin-RevId: 214954192