Make --loading_phase_threads default to the number of CPUs.
Change the behavior of the --loading_phase_threads to match that of --jobs:
default it to a new "auto" magic value that causes Bazel to determine a
reasonable setting (currently the number of CPUs) but allow users to override
it with an explicit count.

This should alleviate threading contention during analysis and reduce the
unresponsiveness of macOS when this happens.  Some tests on large apps show
a reduction of up to 40% (!) on a MacBook Pro with 4 physical cores and a
more modest 15% on a Mac Pro with 6 physical cores.

Note that the previous hack to cap the number of threads to 20 for unit and
integration tests remains.  I'll tackle this separately.

RELNOTES: --loading_phase_threads now defaults to "auto" (not 200, as was
previously the case), which at the moment corresponds to the number of CPUs.
If your sources are on a slow file system, increasing this value may yield
better analysis-time performance when disk caches are cold.
PiperOrigin-RevId: 215151994