Document the repository cache and the --distdir option
Bazel has two standard mechanisms to avoid calling out
to the network unnecessarily. However, so far, the documentation
was hidden in the description of the respective options (and
a few issues and commit messages). Documenting these options
in the context of fetching external documentation makes those
options more visible, and is also the place in the documentation
where users naturally would look for.

PiperOrigin-RevId: 215204872