Rollback of commit 4924566e1c0c0d83a9f4f0ece9f02c098333e666 and changelist 215310545.
RELNOTES: None

*** Reason for rollback ***

Broke bazel CI, see https://buildkite.com/bazel/bazel-bazel/builds/4803

Example:
https://storage.googleapis.com/bazel-buildkite-artifacts/a5047713-623d-4a6b-b2c0-9b58aceac870/src/test/java/com/google/devtools/build/lib/analysis/platform/PlatformAnalysisTests/attempt_1.log

*** Original change description ***

Thread a RootedPath into the Package's BUILD file, versus just a Path. This allows us to use that Root in the common case that the Package is not external, saving the creation of a new Root.

This has a miniscule memory impact, but the more important impact is that we can enumerate all possible Roots now, so Roots can be serialized as constants.

PiperOrigin-RevId: 215361414