Support disabling the repository cache
Support disabling the repository cache in cases where caching is
not necessary, but the additional disk operations are a problem.
This is achieved by interpreting --repository_cache= as a request
to disable the cache; technically, this is an incompatible change,
as currently the empty string is interpreted as "put the cache
top-level into the workspace". However, it is unlikely that this
use of the option is widely used.

Fixes #6229.

Change-Id: I082e814fc36b60b35e6e0ea6c5db8e892f46c4ef
PiperOrigin-RevId: 215365431