Rewrite of the toolchains concept page
This is a total rewrite of the toolchain page rephrasing the existing content in the context of a running example. There is an additional motivation section that contrasts the old non-configurable way of doing things. The presentation is reordered so that we cover how toolchains improve rules, then how we define toolchains, then how we register and trigger them, then how the resolution algorithm works.

Incidentally fixes #6161.

RELNOTES: None
PiperOrigin-RevId: 215561826