Create AnalysisTestResultInfo for in-build testing.
The new object is tied to --experimental_analysis_testing_improvements
This object will eventually be responsible for signaling to Bazel that a test success/failure action should be created on behalf of the current target. This change, however, only exposes this object as a new provider.

Progress toward #6237.

RELNOTES: None.
PiperOrigin-RevId: 215601605