Allow multiple top level target configuration in (a/c)query.
Now when we request targets in the query expression in the target config, we draw the desired configuration from a TopLevelConfigurations object that holds information about top level configurations and the labels that map to them. We store this object instead of the single default configuration we were storing before.

PiperOrigin-RevId: 215605543