Report all known analysis errors to the event bus
This changes --nokeep_going behavior to attempt to report as many known
errors as possible. While --nokeep_going aborts as soon as possible,
multithreading makes it possible for us to receive multiple errors at
this point, so we might as well report all of them.

In the single-error case, we were previously not reporting the error to
the event bus, so the build event protocol was incomplete.

PiperOrigin-RevId: 215828734