file_test: fix for Windows
Rewrite the file_test rule in
@bazel_tools//tools/build_rules:test_rules.bzl as
a macro that combines a genrule with a sh_binary
and a dummy test rule.

Bazel knows the genrule will execute a sh_binary
so it sets up the action's environment correctly
and errors out if Bash is not available.

If the genrule successfully builds its output,
that means the actual assertions passed. The
genrule's output is the input to the downstream
dummy test action, which always passes.

The motivation for this change is that Bazel
didn't use to know (on Windows) that the file_test
rule was executing Bash code. Now with the
sh_binary Bazel knows that it's Bash code.

See https://github.com/bazelbuild/bazel/issues/6122

Change-Id: I159e5dd8f058f8581a4ce7a5f06242bce16374eb