Basic action rewinding test and implementation
Given two genrules, where the 1st produces an output file which is
consumed as an input by the 2nd, if the intermediate file goes missing
before the 2nd genrule has the chance to execute, we'd like to be able
to "rewind" the 1st genrule's action, and make it regenerate its output.

This CL demonstrates that ActionExecutionFunction/SkyframeActionExecutor
support the simple case of rewinding actions for a chain of two genrules.

The action execution node corresponding to the failed action discovers
the set of Skyframe nodes to restart by evaluating, inline, the
dependencies of artifacts associated with its lost inputs. The edges
it adds while doing this are deleted when it restarts.

This is intended as a foundation for future development. Note that
LostInputsExecException is only constructed in tests. Support for
throwing it must still be added to execution strategies.

Future changes will add support for other kinds of actions and artifacts,
and put a limit on how often rewinding is tolerated before failing the
build.

Action rewinding currently depends on Skyframe node restarting, which is
supported only for evaluations which don't track reverse dependencies
between Skyframe nodes.

RELNOTES: None.
PiperOrigin-RevId: 215984081