Allow afterCommand() to modify command exit code
Split the afterCommand() method into two:
 * afterCommand(): can report errors and influence the exit code of a command
 * commandComplete(): called after afterCommand() and can't modify the
   command exit code.

The change for afterCommand() to be able to modify the command exit code
is necessary so that Bazel can block for the BES upload to finish and fail
the command if the upload fails. Additionally, the new commandComplete()
method gives us a safe place to close the command.log as at this point
no new errors can be raised.

RELNOTES: None
PiperOrigin-RevId: 216218312