Querying the remote Capabilities service for remote execution/cache builds.
- call the Capabilities API once per command and save the server capabilities
- check the server API version in order to issue a warning if client uses an unsupported or deprecated version
- check that server supports execution, if remote execution is requested, error otherwise
- check that server digest function matches the client, error otherwise
- check that updating action cache is enabled if requested, error otherwise

Minor refactoring of the RemoteModule initialization logic for clarity. I decided to provide the resulting ServerCapabilities to the remote classes (e.g. RemoteSpawnRunner) on initialization; a different way to do this would be to provide them a handle to the client, and allow lazy fetch. I initially implemented the latter, but switched to the former because it is much simpler.

RELNOTES: query remote server Capabilities (per REAPI v2)
PiperOrigin-RevId: 216353089