Don't let proxy configuration interfere with remote execution.
The existing proxy helper code adjusts JVM properties for setting the
HTTP proxy. This is, based on the source code comments, done to make
jgit use the right proxy.

This code is problematic for two reasons:

- The properties are never cleared between
  ProxyHelper.createProxyIfNeeded() calls, meaning that if a successive
  Git repository were to be part of NO_PROXY, it would still use the
  previously configured proxy.

- Credentials are never cleared, meaning that credentials belonging to
  one proxy may later be sent to the next one.

- The proxy settings end up being used by gRPC for remote execution
  later on, which completely breaks remote builds.

As the new git_repository() no longer uses jgit, remove the code that
sets the JVM properties. It's simply too brittle.