C++: Get linker inputs for cc_fake_binary from CcLinkParams
If we get them from CcLinkParams we can avoid having an extra parameter in
CcLinkingOutputs.

Rolling forward after fixing bug that broke python targets depending on C++ depending on Go. This was caused because we do not always call the buildFinalExtraLinkTimeLibraries() method of CcLinkParams, nor do we have to. By adding back deleted lines from the original the bug is fixed. Also added test.

RELNOTES:none
PiperOrigin-RevId: 216387520