Always execute SymlinkActions that point to absolute paths.
This is needed because we don't know when their target changes and thus we can't invalidate the action graph properly. We work around this by always re-running the symlinking which, as a side effect, makes Bazel check whether its target changed.

This looks inefficient, but it's not *that* inefficient: we only create symlinks to absolute paths for FDO profiles, which is at most twice per build and it's used very rarely anyway (not that I *want* to be able to create absolute symlinks from the exec root, but it was grandfathered in so we have to deal with it...). Dependent actions will not be re-run if the target of the symlink did not change due to change pruning.

RELNOTES: None.
PiperOrigin-RevId: 217452898