Add java_home_runfiles_path to JavaRuntimeInfo
JavaRuntimeInfo provides a runfiles path to bin/java, execpath to the JDK, and execpath to bin/java. This CL adds "java_home_runfiles_path" which provides the runfiles path to the JDK.

Added/modified tests to check all four JavaRuntimeInfo attributes in the java_home cases of absolute, hermetic (prebuilt), and generated.

PiperOrigin-RevId: 217572927