BuildRequestOptions.java: bump the max --jobs setting to keep up with the times.
This was last raised over a year ago, and at least I personally have had two
significant workstation upgrades since then. As far as I can tell 3,000 wasn't
chosen scientifically, so there's no reason to suspect that it's optimal. If it
turns out that the new value is problematic for some reason this can always be
rolled back.

More importantly, the default for --jobs still says the same, so this is purely
an opt in change and is unlikely to be disruptive.

PiperOrigin-RevId: 217617727