Create relative symlinks if the FDO profile is an Artifact (otherwise, the target of the FDO profile symlink will stay an absolute path)
This is useful because symlinks to absolute path are either incorrect or
executed unconditionally and thus cause a lot of Skyframe invalidation.

I *think* we don't even need to create the symlink if the FDO profile is an artifact (we can just use its exec path), but FDO is complex and scary so let's go slowly and make this change an idempotent refactoring (except for the symlink progress message).

RELNOTES: None.
PiperOrigin-RevId: 217672823