Do not fail all the coverage run when there are malformed lines or no coverage.
Some tests will not output any coverage information and we shouldn't fail the coverage command. This is important when we want to compute a combined report after running a batch of tests. In that case if one test doesn't have any coverage information, the combined report will not be created. We want to create the combined report and mark the test as passed.

gcov may also output some malformed line (not often, but it can happen) and we shouldn't fail the whole command because of one line out of thousands.

Fixes #4309.

Closes #6370.

PiperOrigin-RevId: 217701471