Remove support for --incompatible_never_use_embedded_jdk_for_javabase
Fixes #6105

RELNOTES: If the --javabase flag is unset, it Bazel locates a JDK using
the JAVA_HOME environment variable and searching the PATH. If no
JDK is found --javabase will be empty, and builds targeting Java will not
be supported. Previously Bazel would fall back to using the embedded
JDK as a --javabase, but this is no longer default behaviour. A JDK should
be explicitly installed instead to enable Java development
PiperOrigin-RevId: 218388326