testing: add a data-dependency for sync tests
A `bazel sync` requests that all external repositories be fetched. With the
implicit addition of jdk.WORKSPACE to every workspace, this includes a lot
of http archives. Add a data-dependency to tests using `bazel sync` so that
they can be taken from a distdir instead of calling out to the network
for every test.

Change-Id: I8c0dca4f34e11cfead3a419114273254f5addfa3
PiperOrigin-RevId: 218486730