Read the CROSSTOOL from the package of the current cc_toolchain, not from --crosstool_top
This will make the behavior work properly when using platforms, and it will fix
recurring issues when cc_toolchain_suite and cc_toolchain are built as a top
level targets, without setting --crosstool_top (that resulted in trying to
associate cc_toolchain with an unrelated CROSSTOOL file and fail the build).
This could also happen when using `bazel cquery` or `bazel aquery`.

RELNOTES: CROSSTOOL file is now read from the package of cc_toolchain, not from
the package of cc_toolchain_suite. This is not expected to break anybody since
cc_toolchain_suite and cc_toolchain are commonly in the same package.
PiperOrigin-RevId: 218516709