Merge new_file overrides into a single method.
This one method represents the only occurrence of @SkylarkCallable method overriding in the BUILD API. With removal of overrides, we can do a variety of cleanup of interpreter code.

The new_file method is deprecated, so the added signature-matching complexity for this method is a small price to pay for the subsequent cleanup. Previously, only *some* of the overrides of this method were restricted subject to --incompatible_new_actions_api. This change restricts *all* such overrides.

Progress toward #6490.

RELNOTES:
  All overrides of Starlark's ctx.new_file function are now deprecated.
  Try the `--incompatible_new_actions_api` flag to ensure your code is forward-compatible.
PiperOrigin-RevId: 218526541