For source directory action inputs, depend on all files in the subtree
This fixes a number of long-standing correctness issues where Bazel did not
notice if files within the purview of an action changed.

Consider this BUILD rule:
genrule(
  name = "bad",
  outs = ["ls.txt"],
  srcs = ["dir"],
  cmd = "ls -l test/dir > $@",
)

If "dir" is a directory, then Bazel does not notice changes to files
underneath dir, and also does not give an error message. It only emits a
warning like this:
"WARNING: /path/to/BUILD:1:1: input 'test/dir' to //test:bad is a directory; dependency checking of directories is unsound"

We use a startup flag for rollout since the new code requires additional
Skyframe dependencies. Use like this:
bazel --host_jvm_args=-DBAZEL_TRACK_SOURCE_DIRECTORIES=1 build ...

With this change, Bazel assumes that all the files underneath dir/ are action
inputs (i.e., it implicitly globs all the files) and reruns the action
whenever any of these files change. The main differences to glob(["dir/**"])
are:
- the file names do not have to be valid labels; the label syntax currently
  restricts the allowed characters to a small subset of UTF-8 (mainly ASCII)
- the files cannot be directly referenced from another package unless they
  are also explicitly mentioned in the package
- globbing happens during execution, not loading; this can have a significant
  effect on performance for packages that have a lot of globs / directory
  references - globs are evaluated eagerly during loading, even if only a
  single rule from that package is needed

This change does not affect remote execution, which continues to not allow
such inputs, although this could be changed subsequently.

We may in the future require that directory references in BUILD files end with
a trailing forward slash ('/') to indicate this fact to a reader of the BUILD
file.

PiperOrigin-RevId: 218817101