external repositories: be explicit about restarts in fetching
When reporting progress about fetching external repositories, be
very explicit about restarts. Especially, as restarts are rare
now that we prefetch everything, we should not hide this situation
in the UI. Also, report fetching an external repository finished
as soon as the rule is completed; the book-keeping with marker
files etc is bazel internal.

Change-Id: I55ab0cbd15ad92bc0fff81020bf81d94f8c6771d
PiperOrigin-RevId: 219104347