Make ijar keep types/member functions that have annotated with @KeepForCompile.
The inline Kotlin language feature causes method implementations to be a part of the jar's compiletime interface. We have a Kotlin compiler plugin that will annotate inlined methods with @KeepForCompile and ijar can use this to include code that is a part of the library's compiletime interface into the stripped jar.

This iteration keeps the entire class, in the future it only needs to keep the
method implementation and any constant pool entries the method accesses.

PiperOrigin-RevId: 219553768